feat: change identityDocument to long

+++ src/main/java/com/powerup/user/users/application/dto/request/CreateUserRequest.java
@@ -4,6 +4,6 @@ package com.powerup.user.users.application.dto.request;
 
 import java.time.LocalDate;
 
-public record CreateUserRequest(String firstName, String lastName, Integer identityDocument, String phoneNumber,
+public record CreateUserRequest(String firstName, String lastName, Long identityDocument, String phoneNumber,
                                 LocalDate birthDate, String email, String password, Long role) {
 }
